+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # CHANGELOG
 
+## 0.1.11
+- added Function to delete a directory
+
 ## 0.1.10 - 05.10.2025
 - added Function to Check the OS Type to make code in the script which runs only
 on a particular OS